Use snapshots for in-memory copies of layered worldstate (#4700)
* use snapshots for in-memory copies of layered worldstate
* ensure we close snapshot storage used for layered world state copies
* wrap access to CachedSnapshotWorldState in Optional.ofNullable to prevent NPE when trying to close from AbstractTrieLogManager
